now int's in Expression.java and Expression3.java

+++ Expression.java
@@ -3,5 +3,5 @@
  * @version $$Id$$
  */
 public interface Expression {
-    double evaluate(double x);
+    int evaluate(int x);
 }
